+++ src/main/java/com/shadorc/shadbot/core/command/CommandManager.java
@@ -11,7 +11,7 @@ import com.shadorc.shadbot.command.currency.CoinsCmd;
 import com.shadorc.shadbot.command.currency.LeaderboardCmd;
 import com.shadorc.shadbot.command.currency.TransferCoinsCmd;
 import com.shadorc.shadbot.command.fun.*;
-import com.shadorc.shadbot.command.game.LotteryCmd;
+import com.shadorc.shadbot.command.game.lottery.LotteryCmd;
 import com.shadorc.shadbot.command.game.RussianRouletteCmd;
 import com.shadorc.shadbot.command.game.blackjack.BlackjackCmd;
 import com.shadorc.shadbot.command.game.dice.DiceCmd;
